nvim: move toggleterm configuration to the after/plugin directory
